# 6-full-api/docs/swagger.yaml
basePath: /
definitions:
  dto.CreateProductInput:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  dto.CreateUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.GetJWTInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.GetJWTOutput:
    properties:
      access_token:
        type: string
    type: object
  entity.Product:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    name: Stefano Kaefer
  description: Product API developed on Booster program
  title: Product API Booster
  version: "0.1"
paths:
  /products:
    post:
      consumes:
      - application/json
      description: Create product in the api
      parameters:
      - description: Create product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Product'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create product
      tags:
      - products
  /users:
    post:
      consumes:
      - application/json
      description: Create User in the api
      parameters:
      - description: Create User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.User'
        "500":
          description: Internal Server Error
      summary: Create User
      tags:
      - users
  /users/token:
    post:
      consumes:
      - application/json
      description: Get Token JWT
      parameters:
      - description: Get JWT
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetJWTInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetJWTOutput'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get JWT
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"